+++ src/test/resources/inspections/CollectFunctionInCollectionInspection/nonMatches/collect_on_no_type_variable.php
@@ -2,4 +2,4 @@
 
 function($variable) {
     collect($variable);
-}
+};